Add JavaDoc comments to Task class
