+++ build.gradle
@@ -28,11 +28,11 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("khat.Khat")
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "khat"
     archiveClassifier = null
 }
